Replace Entity GUIDs with ID Remapping on Load (#69)
- Changing how entities are stored yet again! We previously changed from using entity IDs that contains index and version information and replaced those with a 128-bit random integer, allowing for persistent IDs across saving. This worked fine but bloated the entities.
- Now, each entity is represented by a 64bit unsigned int called `ecs::Identifier`. This is an `enum class` and should be used as an opaque identifier.
- Component modifying functions are added back to the `ecs::Registry`. `ecs::Entity` is no longer a friend class since it is just a light wrapper around an identifier and calls the new functions on the reg.
- Really, we don't care about about IDs being persistent. The main motivation was to enable having parent/child relationships between entities, but the actual values of the IDs are not important. We just need to make sure that there is an isomorphism between IDs on two separate runs.
- In `Loader`, when entities are loaded, do a first pass to create the new IDs and keep an `std::unordered_map` from old to new IDs.
- When loading components, if an attribute is an `ecs::Identifier` (which is now just a `std::uint64_t`), transform it via the id mapping. Do the same when copying scenes. With this, IDs may change between different runs, but `ecs::Identifier`s on components will still map to the correct entity.
- With the change, add `ParentComponent` which stores an identifier. This is currently unused in the engine.
- Remove a few use cases of the `(ev::Event& even, auto&& data)` subscribe function on the `ev::Dispatcher` as it caused some segfaults. Need to figure out why.
- Integrate `ecs::Identifier` into many places. The intention is that most places that use `ecs::Entity` will be replaced. It is currently represents as userdata in Lua just like an `ecs::Entity`, however in the future we should change this to just be an integer.

+++ Sprocket/Utility/SparseSet.h
@@ -11,7 +11,7 @@
 
 namespace Sprocket {
 
-template <typename ValueType, typename IndexType = std::size_t>
+template <typename ValueType, typename IndexType = std::uint32_t>
 class SparseSet
 {
 public:
